# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 绿化管理系统 (GMS) API
  version: 1.0.0
  description: |
    绿化管理系统 RESTful API，管理绿化区块、植物品种、种植记录和养护记录。

    ## 认证方式
    登录后获取 JWT Token，在请求 Header 中携带：
    ```
    Authorization: Bearer <token>
    ```

    ## 通用响应格式
    ```json
    {
      "code": 200,
      "message": "成功",
      "data": {}
    }
    ```

    ## 权限说明
    - `/admin/` 前缀：管理员接口
    - 其余：用户端接口

servers:
  - url: http://localhost:8080
    description: 开发环境

tags:
  - name: 认证
    description: 用户登录、注册、登出
  - name: 用户管理
    description: 管理员对用户的 CRUD 操作
  - name: 区块管理
    description: 绿化区块的增删改查
  - name: 植物品种
    description: 植物品种管理
  - name: 种植记录
    description: 种植记录的增删改查
  - name: 养护记录
    description: 养护记录的增删改查
  - name: 统计
    description: 首页统计数据

paths:
  /auth/user/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /auth/admin/login:
    post:
      tags: [认证]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, phone]
              properties:
                username:
                  type: string
                password:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: 注册成功

  /auth/logout:
    post:
      tags: [认证]
      summary: 用户登出
      responses:
        '200':
          description: 登出成功

  /admin/users:
    get:
      tags: [用户管理]
      summary: 获取用户列表（分页）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'

  /admin/users/{id}:
    get:
      tags: [用户管理]
      summary: 获取用户详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [用户管理]
      summary: 更新用户信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [用户管理]
      summary: 删除用户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功

  /areas:
    get:
      tags: [区块管理]
      summary: 获取区块列表（分页）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: areaName
          in: query
          schema:
            type: string
          description: 区块名称（模糊搜索）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAreas'

  /areas/{areaId}:
    get:
      tags: [区块管理]
      summary: 获取区块详情
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'

  /admin/areas:
    post:
      tags: [区块管理]
      summary: 新增区块
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaCreate'
      responses:
        '200':
          description: 新增成功

  /admin/areas/{areaId}:
    put:
      tags: [区块管理]
      summary: 更新区块信息
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AreaUpdate'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [区块管理]
      summary: 删除区块
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功

  /plants/species:
    get:
      tags: [植物品种]
      summary: 获取植物品种列表（分页）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: speciesName
          in: query
          schema:
            type: string
          description: 品种名称（模糊搜索）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSpecies'

  /admin/plants/species:
    post:
      tags: [植物品种]
      summary: 新增植物品种
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeciesCreate'
      responses:
        '200':
          description: 新增成功

  /planting/records:
    get:
      tags: [种植记录]
      summary: 获取种植记录列表（分页）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: speciesId
          in: query
          schema:
            type: string
          description: 品种编号（可选）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPlantingRecords'
    post:
      tags: [种植记录]
      summary: 新增种植记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantingRecordCreate'
      responses:
        '200':
          description: 新增成功

  /planting/records/{id}:
    put:
      tags: [种植记录]
      summary: 更新种植记录
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantingRecordCreate'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [种植记录]
      summary: 删除种植记录
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功

  /maintenance/records:
    get:
      tags: [养护记录]
      summary: 获取养护记录列表（分页）
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: speciesId
          in: query
          schema:
            type: string
          description: 品种编号（可选）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMaintenanceRecords'
    post:
      tags: [养护记录]
      summary: 新增养护记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRecordCreate'
      responses:
        '200':
          description: 新增成功

  /maintenance/records/{id}:
    put:
      tags: [养护记录]
      summary: 更新养护记录
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRecordCreate'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [养护记录]
      summary: 删除养护记录
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功

  /stats/dashboard:
    get:
      tags: [统计]
      summary: 获取首页统计数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'

  /stats/recent-activities:
    get:
      tags: [统计]
      summary: 获取近日活动
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentActivities'

  /stats/plant-distribution:
    get:
      tags: [统计]
      summary: 获取植物分布数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDistribution'

components:
  schemas:
    TokenResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "成功"
        data:
          type: object
          properties:
            token:
              type: string

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        registerTime:
          type: string
          format: date-time
        maintenanceCount:
          type: integer
        plantingCount:
          type: integer

    UserUpdate:
      type: object
      properties:
        username:
          type: string
        phone:
          type: string
        role:
          type: string
        maintenanceCount:
          type: integer
        plantingCount:
          type: integer

    PaginatedUsers:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/User'
            totalPages:
              type: integer
            currentPage:
              type: integer
            hasNext:
              type: boolean
            hasPrevious:
              type: boolean

    Area:
      type: object
      properties:
        areaId:
          type: string
        areaName:
          type: string
        areaSize:
          type: number
        locationDescription:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedTime:
          type: string
          format: date-time

    AreaCreate:
      type: object
      required: [areaId, areaName, areaSize]
      properties:
        areaId:
          type: string
        areaName:
          type: string
        areaSize:
          type: number
        locationDescription:
          type: string

    AreaUpdate:
      type: object
      properties:
        areaName:
          type: string
        areaSize:
          type: number
        locationDescription:
          type: string

    PaginatedAreas:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/Area'
            totalPages:
              type: integer
            currentPage:
              type: integer
            hasNext:
              type: boolean
            hasPrevious:
              type: boolean

    Species:
      type: object
      properties:
        speciesId:
          type: string
        speciesName:
          type: string
        createdTime:
          type: string
          format: date-time

    SpeciesCreate:
      type: object
      required: [speciesId, speciesName]
      properties:
        speciesId:
          type: string
        speciesName:
          type: string

    PaginatedSpecies:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/Species'
            totalPages:
              type: integer
            currentPage:
              type: integer
            hasNext:
              type: boolean
            hasPrevious:
              type: boolean

    PlantingRecordCreate:
      type: object
      required: [areaId, speciesId, plantingTime, plantCount, responsiblePerson]
      properties:
        areaId:
          type: string
        speciesId:
          type: string
        plantingTime:
          type: string
          format: date-time
        plantCount:
          type: integer
        responsiblePerson:
          type: string
        description:
          type: string

    PaginatedPlantingRecords:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            records:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  areaId:
                    type: string
                  speciesId:
                    type: string
                  plantingTime:
                    type: string
                    format: date-time
                  plantCount:
                    type: integer
                  responsiblePerson:
                    type: string
                  description:
                    type: string
                  createdTime:
                    type: string
                    format: date-time
                  areaName:
                    type: string
                  speciesName:
                    type: string
            totalPages:
              type: integer
            currentPage:
              type: integer
            hasNext:
              type: boolean
            hasPrevious:
              type: boolean

    MaintenanceRecordCreate:
      type: object
      required: [areaId, speciesId, maintenanceTime, plantCount, maintainer]
      properties:
        areaId:
          type: string
        speciesId:
          type: string
        maintenanceTime:
          type: string
          format: date-time
        plantCount:
          type: integer
        maintainer:
          type: string
        description:
          type: string

    PaginatedMaintenanceRecords:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            records:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  areaId:
                    type: string
                  speciesId:
                    type: string
                  maintenanceTime:
                    type: string
                    format: date-time
                  plantCount:
                    type: integer
                  maintainer:
                    type: string
                  description:
                    type: string
                  createdTime:
                    type: string
                    format: date-time
                  areaName:
                    type: string
                  speciesName:
                    type: string
            totalPages:
              type: integer
            currentPage:
              type: integer
            hasNext:
              type: boolean
            hasPrevious:
              type: boolean

    DashboardStats:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            areaCount:
              type: integer
            monthlyMaintenanceCount:
              type: integer
            maintainerCount:
              type: integer
            planterPercentage:
              type: number
            maintainerPercentage:
              type: number
            plantSpeciesCount:
              type: integer
            planterCount:
              type: integer
            monthlyPlantingCount:
              type: integer

    RecentActivities:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            plantingActivities:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  areaId:
                    type: string
                  areaName:
                    type: string
                  speciesId:
                    type: string
                  speciesName:
                    type: string
                  responsiblePerson:
                    type: string
                  plantingTime:
                    type: string
                    format: date-time
                  type:
                    type: string
            maintenanceActivities:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  areaId:
                    type: string
                  areaName:
                    type: string
                  speciesId:
                    type: string
                  speciesName:
                    type: string
                  maintainer:
                    type: string
                  maintenanceTime:
                    type: string
                    format: date-time
                  type:
                    type: string

    PlantDistribution:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              areaId:
                type: string
              areaName:
                type: string
              speciesCount:
                type: integer
              plantTotalCount:
                type: integer

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT